Set the default value of notifications_enable to false
<soapbox>
This is just a demo so it doesn't really matter, but it drives me crazy
when an app opts me in by default to any notifications at all. All
notifications should default to off and require an explicit opt-in
from the user!

The example settings activity in Android Studio should also make the
default value false. Making the default true gives developers the
impression that it's ok to send notifications to users without their
approval.
</soapbox>